chore: rename folder with typed redux hooks
