+++ src/components/Lists/Tasks.tsx
@@ -2,7 +2,7 @@
 
 /* APPLICATION */
 import { selectAllTasks } from '../../store/tasksSlice';
-import { useAppSelector } from '../../hooks/hooks';
+import { useAppSelector } from '../../hooks/reduxTypedHooks';
 import { ListItem } from './ListItem';
 
 export const Tasks = () => {
