+++ src/components/Modal/ModalCreateItem.tsx
@@ -5,7 +5,7 @@ import React, { useState } from 'react';
 import { categoriesAdded } from '../../store/categoriesSlice';
 import { tasksAdded } from '../../store/tasksSlice';
 import { useCheckPath } from '../../hooks/useCheckPath';
-import { useAppDispatch } from '../../hooks/hooks';
+import { useAppDispatch } from '../../hooks/reduxTypedHooks';
 import { Modal } from './Modal';
 import { ModalHeader } from './ModalHeader';
 import { ModalInput } from './ModalInput';
