+++ src/components/Modal/ModalRemoveItem.tsx
@@ -5,7 +5,7 @@ import React from 'react';
 import { categoriesRemoved } from '../../store/categoriesSlice';
 import { tasksRemoved, tasksClearedCategories } from '../../store/tasksSlice';
 import { useCheckPath } from '../../hooks/useCheckPath';
-import { useAppDispatch } from '../../hooks/hooks';
+import { useAppDispatch } from '../../hooks/reduxTypedHooks';
 import { Modal } from './Modal';
 import { ModalHeader } from './ModalHeader';
 import { ModalFooter } from './ModalFooter';
